Prolog - Assignment 1
Question 1:
Use a single cut in the program such that only one solution for each X
is found. For example, if X =
a, Y = b is found, the solution X = a, Y = c should be excluded.

Question 2:
Develop rules for finding numbers which cubic is below a certain number.
a) Create the rule cubeLess(X,B,R) which calculates the power of 3 and
the remainder such that
b =x3+r. Therefore, for a query cubeLess( 2, 10, R) , the result should
be R=2. Note, that you will
not be able to call the rule without X and B instantiated.
b) Use the rule cubeLess from a) to find all numbers with a cube smaller
than a given limit by
creating a new rule smallerCube(X,B). Print your results to the console
with write/1 and
writeln/1 . Example:
?- smallerCube(X,130).
1 rest 129
2 rest 122
3 rest 103
4 rest 66
5 rest 5
Because 13= 1 + 129 = 130, 23= 8 + 122 = 130 and so on.
c) Create another (and separate) rule restSum(B,S)to add up all the
remainders from finding all
cubes up to an upper limit.
?- restSum(130,S).
S = 425.
d) Write another rule to print all the rest sums that are multiples of 3
within a range. For example,
in the range from 1 to 20:
?- showAllRestSum(1,20).
1 rest 0
4 rest 3
7 rest 6
9 rest 9
12 rest 15
15 rest 21
18 rest 27
true.